+++ packages/ui-shared/src/components/BenefitsIcons/BenefitsIconsTile.tsx
@@ -1,0 +1,43 @@
+import * as React from 'react';
+import PropTypes from 'prop-types';
+import convert from 'htmr';
+import './style/BenefitsIconsTile.less';
+import { Header, Paragraph, cnCreate } from '@megafon/ui-core';
+import { IBenefit, IconPositionEnum } from './types';
+
+export interface IBenefitsIconsTile extends IBenefit {
+    iconPosition?: IconPositionEnum;
+}
+
+const cn = cnCreate('benefits-icons-tile');
+const BenefitsIconsTile: React.FC<IBenefitsIconsTile> = ({
+    title,
+    text,
+    icon,
+    iconPosition = IconPositionEnum.LEFT_TOP,
+}) => (
+    <div className={cn('', { 'icon-position': iconPosition })}>
+        <div className={cn('svg-icon')}>{icon}</div>
+        <div className={cn('content-wrapper')}>
+            {title && (
+                <Header className={cn('title')} as="h5">
+                    {convert(title)}
+                </Header>
+            )}
+            {text && (
+                <Paragraph className={cn('text')} hasMargin={false}>
+                    {convert(text)}
+                </Paragraph>
+            )}
+        </div>
+    </div>
+);
+
+BenefitsIconsTile.propTypes = {
+    title: PropTypes.string,
+    text: PropTypes.string,
+    iconPosition: PropTypes.oneOf(Object.values(IconPositionEnum)),
+    icon: PropTypes.node,
+};
+
+export default BenefitsIconsTile;