+++ packages/ui-shared/CHANGELOG.md
@@ -4,6 +4,7 @@
 - **Constructor**
     - create ButtonLinkBox component
     - create TitleDescriptionBox component
+    - create BenefitIcons component
 
 ## Release 1.0.30
 - **package.json**